% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_info.R
\name{gdal_build_info}
\alias{gdal_build_info}
\title{Runtime GDAL build flags}
\usage{
gdal_build_info()
}
\value{
named \code{list} of flags set when building linked GDAL library.
}
\description{
Runtime GDAL build flags
}
